Fix issue when updating a free access plan

+++ includes/server/class-llms-rest-access-plans-controller.php
@@ -5,7 +5,7 @@
  * @package LifterLMS_REST/Classes/Controllers
  *
  * @since 1.0.0-beta.18
- * @version 1.0.0-beta.20
+ * @version [version]
  */
 
 defined( 'ABSPATH' ) || exit;
@@ -611,6 +611,7 @@ class LLMS_REST_Access_Plans_Controller extends LLMS_REST_Posts_Controller {
 	 * These properties must be compared to the saved value before updating, because if equal they will produce an error(see update_post_meta()).
 	 *
 	 * @since 1.0.0-beta.18
+	 * @since [version] Cast `price` property to float.
 	 *
 	 * @param array $to_set      Array of properties to be set.
 	 * @param array $saved_props Array of LLMS_Access_Plan properties as saved in the db.
@@ -636,7 +637,7 @@ class LLMS_REST_Access_Plans_Controller extends LLMS_REST_Posts_Controller {
 		} else {
 
 			$subordinate_props['is_free']     = 'yes';
-			$subordinate_props['price']       = 0;
+			$subordinate_props['price']       = (float) 0;
 			$subordinate_props['frequency']   = 0;
 			$subordinate_props['on_sale']     = 'no';
 			$subordinate_props['trial_offer'] = 'no';
